requirements.txt
django-bootstrap-v5
django-admin-interface
django-crispy-forms
xhtml2pdf
django-crispy-forms 
crispy-bootstrap5
pillow
passlib
